cleaning up comments, partial files
