+++ src/FancyMouse.Common/Bezels/BezelPrimitives.cs
@@ -6,8 +6,8 @@ internal static class BezelPrimitives
 {
     /// <summary>
     /// Implements a cosine easing function that can be used to create a
-    /// smooth transition between two values on a continuous axis
-    /// within an interval.
+    /// smooth transition between two values on a continuous axis within
+    /// an interval.
     /// <code>
     ///  +-----+-----+-----+
     ///   -----._    .       - start value
