+++ src/FancyMouse.Common/Bezels/BezelProfileCurved.cs
@@ -8,12 +8,12 @@ namespace FancyMouse.Common.Bezels;
 //
 // Cross-section layout (position 0 = outer arc edge, n = content boundary):
 //
-//   position:  0      d           n-d      n
-//              │      │            │       │
-//   θ (normal):0 ──→ π/2 ──────── π/2 ──→ π
-//   cos(θ):   +1 ──→  0  ──────── 0  ──→ -1
-//              │outer ring│  flat  │inner ring│
-//              │(highlight)│ (none) │(shadow) │
+//   position:   0      d           n-d     n
+//               │      │            │      │
+//   θ (normal): 0 ──→ π/2 ──────── π/2 ──→  π
+//   cos(θ):    +1 ──→  0  ────────  0  ──→ -1
+//               │outer ring│  flat  │inner ring│
+//               │(highlight)│ (none) │(shadow) │
 //
 // GetProfileNormal maps position to a normal angle; the shared helpers in
 // BezelProfile convert that angle to highlight / shadow intensities.
