+++ src/FancyMouse.Common/Blurring/ScreenshotBlurState.cs
@@ -3,9 +3,9 @@ using System.Drawing;
 namespace FancyMouse.Common.Blurring;
 
 /// <summary>
-/// One physical screen's blur state within <see cref="ScreenshotBlurPipeline"/> - see its remarks for what
-/// "todo"/"doing" (<see cref="BlurInProgress"/>)/"done" mean. Internal - <see cref="ScreenshotBlurPipeline"/>
-/// is the only thing that ever touches an instance of this, always under its own lock.
+/// Represents the state of the blurred screenshot images for one physical screen
+/// within a <see cref="ScreenshotBlurPipeline"/> - see its remarks for what
+/// "todo" / "doing" (<see cref="BlurInProgress"/>) / "done" mean.
 /// </summary>
 internal sealed class ScreenshotBlurState
 {
