+++ src/FancyMouse.Common/Interop/Win32WindowProc.cs
@@ -1,5 +1,4 @@
-﻿using Windows.Win32;
-using Windows.Win32.Foundation;
+﻿using Windows.Win32.Foundation;
 using Windows.Win32.UI.WindowsAndMessaging;
 
 namespace FancyMouse.Common.Interop;
