+++ src/FancyMouse.WinUI3/NLog.config
@@ -3,6 +3,8 @@
       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
 
   <targets async="true">
+    <!-- fileName below is a fallback only - App.InitializeLogger overrides it at startup with a
+         timestamp shared with the telemetry file, so the two can be matched up by name. -->
     <target name="logfile"
             xsi:type="File"
             fileName="${specialfolder:folder=LocalApplicationData}/FancyMouse/Logs/applog_${processinfo:property=StartTime:format=yyyy-MM-dd_HH_mm_ss}.txt"
