+++ src/FancyMouse.WinUI3/UI/PreviewPane.xaml.cs
@@ -1,26 +1,9 @@
-using System.Drawing;
-using System.Drawing.Imaging;
-using System.Runtime.InteropServices;
-using System.Runtime.InteropServices.WindowsRuntime;
-
 using FancyMouse.Common.Blurring;
-using FancyMouse.Common.Helpers;
 using FancyMouse.Models.Display;
-using FancyMouse.Models.Drawing;
 using FancyMouse.Models.Layout;
-using FancyMouse.Models.Styles;
 
-using Microsoft.UI.Input;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml.Input;
-using Microsoft.UI.Xaml.Media;
-using Microsoft.UI.Xaml.Media.Animation;
-using Microsoft.UI.Xaml.Media.Imaging;
-
-using Windows.System;
-
-using Image = Microsoft.UI.Xaml.Controls.Image;
 
 namespace FancyMouse.WinUI3.UI;
 
@@ -115,461 +98,4 @@ public sealed partial class PreviewPane : UserControl
     /// without moving the pointer anywhere.
     /// </summary>
     public event EventHandler? Cancel;
-
-    /// <summary>
-    /// Backfills the real screenshot for a single screen, replacing its placeholder fill -
-    /// called by the hosting window once its capture pipeline has produced the image for that
-    /// screen. Takes ownership of <paramref name="image"/> - see <see cref="IScreenshotCaptureSink.SetScreenshotAsync"/>.
-    /// If <paramref name="screenLayout"/> isn't part of the current <see cref="Layout"/> any more
-    /// (e.g. a newer activation replaced it while this capture was still in flight), the call
-    /// disposes <paramref name="image"/> and returns rather than throwing - a stale, superseded
-    /// capture result has nowhere left to go.
-    /// </summary>
-    /// <remarks>
-    /// Also hands <paramref name="image"/> to <see cref="blurPipeline"/>, which generates and
-    /// retains a blurred stand-in from it for <see cref="GetHistoricalPlaceholder"/> to use as a
-    /// *future* activation's initial placeholder while its own fresh screenshot is still
-    /// loading - see <see cref="ScreenshotBlurPipeline"/> for how that's kept off the UI thread and how
-    /// stale/superseded work gets discarded.
-    /// </remarks>
-    public void SetScreenshot(ScreenLayout screenLayout, Bitmap image)
-    {
-        ArgumentNullException.ThrowIfNull(screenLayout);
-        ArgumentNullException.ThrowIfNull(image);
-
-        var index = this.screenSlots.FindIndex(
-            s => object.ReferenceEquals(s.ScreenLayout, screenLayout));
-        if (index < 0)
-        {
-            image.Dispose();
-            return;
-        }
-
-        this.CrossfadeContent(this.screenSlots[index], PreviewPane.ToBitmapImage(image));
-        this.blurPipeline.SetScreenshot(screenLayout.ScreenInfo, image);
-    }
-
-    /// <summary>
-    /// Fades <paramref name="newSource"/> in over <paramref name="slot"/>'s existing content, rather
-    /// than swapping <see cref="ScreenSlot.ContentImage"/>'s own <c>Source</c> instantly - WinUI
-    /// has no built-in way to crossfade an <see cref="Image"/> between two sources, so this adds
-    /// a temporary overlay <see cref="Image"/> at the same position, animates its
-    /// <see cref="UIElement.Opacity"/> from 0 to 1, then (once that's done) makes
-    /// <paramref name="newSource"/> the slot's own real content and discards the overlay - by
-    /// then it's fully opaque, so the swap itself is invisible.
-    /// </summary>
-    private void CrossfadeContent(ScreenSlot slot, ImageSource newSource)
-    {
-        var overlay = new Image
-        {
-            Source = newSource,
-            Stretch = Stretch.Fill,
-            Opacity = 0,
-            Width = slot.ContentImage.Width,
-            Height = slot.ContentImage.Height,
-        };
-        Canvas.SetLeft(overlay, Canvas.GetLeft(slot.ContentImage));
-        Canvas.SetTop(overlay, Canvas.GetTop(slot.ContentImage));
-        this.ScreensCanvas.Children.Add(overlay);
-
-        var animation = new DoubleAnimation
-        {
-            From = 0,
-            To = 1,
-            Duration = new Duration(PreviewPane.ContentTransitionDuration),
-            EasingFunction = new QuadraticEase { EasingMode = EasingMode.EaseOut },
-        };
-        Storyboard.SetTarget(animation, overlay);
-        Storyboard.SetTargetProperty(animation, "Opacity");
-
-        var storyboard = new Storyboard();
-        storyboard.Children.Add(animation);
-        storyboard.Completed += (_, _) =>
-        {
-            slot.ContentImage.Source = newSource;
-            this.ScreensCanvas.Children.Remove(overlay);
-        };
-        storyboard.Begin();
-    }
-
-    /// <summary>
-    /// Returns this screen's last real screenshot, blurred, if <see cref="blurPipeline"/> still
-    /// has one available - or <see langword="null"/> (falling back to the plain placeholder
-    /// fill) if not - for a screen whose current activation hasn't captured yet. See
-    /// <see cref="ScreenshotBlurPipeline.TryGet"/> for the freshness rules.
-    /// </summary>
-    private WriteableBitmap? GetHistoricalPlaceholder(ScreenInfo screenInfo)
-    {
-        WriteableBitmap? result = null;
-        this.blurPipeline.TryGet(screenInfo, blurredImage => result = PreviewPane.ToBitmapImage(blurredImage));
-        return result;
-    }
-
-    private static void OnLayoutChanged(DependencyObject sender, DependencyPropertyChangedEventArgs e)
-        => ((PreviewPane)sender).ApplyLayout((PreviewLayout?)e.NewValue);
-
-    private void ApplyLayout(PreviewLayout? layout)
-    {
-        if (layout is null)
-        {
-            this.ScreensCanvas.Children.Clear();
-            this.screenSlots = [];
-            this.Width = 0;
-            this.Height = 0;
-            this.BackgroundImage.Source = null;
-            return;
-        }
-
-        var scale = this.GetRasterizationScale();
-        var bounds = layout.CanvasLayout.CanvasBounds.OuterBounds;
-        this.Width = (double)bounds.Width / scale;
-        this.Height = (double)bounds.Height / scale;
-
-        using var backgroundBitmap = DrawingHelper.RenderBackground(layout.CanvasLayout);
-        this.BackgroundImage.Source = PreviewPane.ToBitmapImage(backgroundBitmap);
-
-        var newScreenLayouts = layout.CanvasLayout.DeviceLayouts
-            .SelectMany(deviceLayout => deviceLayout.ScreenLayouts)
-            .ToList();
-
-        // a new activation - tell blurPipeline which physical screens currently exist, before
-        // consulting it for placeholders below, so only blurs for screens that are actually
-        // still connected ever get used (see ScreenshotBlurPipeline.SetActiveScreens)
-        this.blurPipeline.SetActiveScreens(
-            newScreenLayouts.Select(screenLayout => screenLayout.ScreenInfo).ToList());
-
-        var newSlots = new List<ScreenSlot>(newScreenLayouts.Count);
-        for (var i = 0; i < newScreenLayouts.Count; i++)
-        {
-            var screenLayout = newScreenLayouts[i];
-            var previousSlot = (i < this.screenSlots.Count) ? this.screenSlots[i] : null;
-
-            if (previousSlot is not null && PreviewPane.CanReuse(previousSlot.ScreenLayout, screenLayout))
-            {
-                // pixel-identical bezel/placeholder to what's already on screen at this
-                // position - keep the visuals, just point the slot at the new ScreenLayout
-                // instance (SetScreenshot looks slots up by reference) and clear the content
-                // image (falling back to this screen's blurred stand-in, if any - see
-                // GetHistoricalPlaceholder) since the screenshot it was showing is now stale
-                // desktop state regardless of whether the bezel itself changed
-                previousSlot.ContentImage.Source = this.GetHistoricalPlaceholder(screenLayout.ScreenInfo);
-
-                // CanReuse only guarantees the *physical-pixel* bounds are unchanged - the
-                // DIP-space Width/Height/Canvas.Left/Top these elements were last positioned
-                // at also depend on scale (see GetRasterizationScale), which can differ from
-                // last time if this activation's window landed on a different-DPI monitor
-                // than the previous one. Reapplying position/size here is cheap (a handful of
-                // property writes, no re-render) and keeps that in sync even though the
-                // visuals themselves are being reused as-is.
-                var screenBounds = screenLayout.ScreenBounds;
-                PreviewPane.PositionElement(previousSlot.BezelImage, screenBounds.OuterBounds, scale);
-                if (previousSlot.PlaceholderRectangle is not null)
-                {
-                    PreviewPane.PositionElement(previousSlot.PlaceholderRectangle, screenBounds.PaddingBounds, scale);
-                }
-
-                PreviewPane.PositionElement(previousSlot.ContentImage, screenBounds.ContentBounds, scale);
-
-                newSlots.Add(new ScreenSlot(
-                    screenLayout, previousSlot.BezelImage, previousSlot.PlaceholderRectangle, previousSlot.ContentImage));
-            }
-            else
-            {
-                this.RemoveScreenSlot(previousSlot);
-                newSlots.Add(this.CreateScreenSlot(screenLayout, scale));
-            }
-        }
-
-        // trim any slots left over from a layout that had more screens than this one does
-        // (e.g. a monitor was unplugged) - everything up to newScreenLayouts.Count was
-        // already handled (reused or replaced) above
-        for (var i = newScreenLayouts.Count; i < this.screenSlots.Count; i++)
-        {
-            this.RemoveScreenSlot(this.screenSlots[i]);
-        }
-
-        this.screenSlots = newSlots;
-    }
-
-    /// <summary>
-    /// Reports whether <paramref name="current"/> would produce a pixel-identical bezel and
-    /// placeholder to <paramref name="previous"/> - if so, <see cref="ApplyLayout"/> reuses the
-    /// existing visuals instead of re-rendering them. Deliberately narrow: only the fields that
-    /// actually feed <see cref="CreateScreenSlot"/> are compared (the three bounds rectangles
-    /// used for positioning, the border style used for the bezel, and the background color used
-    /// for the placeholder) - margin/padding style and the rest of <see cref="BoxStyle"/> only
-    /// matter insofar as they already show up in the bounds, so comparing bounds directly covers
-    /// them without needing to compare every style field individually.
-    /// </summary>
-    private static bool CanReuse(ScreenLayout previous, ScreenLayout current)
-        => PreviewPane.CanReuse(previous.ScreenBounds, current.ScreenBounds)
-        && PreviewPane.CanReuse(previous.ScreenStyle.BorderStyle, current.ScreenStyle.BorderStyle)
-        && previous.ScreenStyle.BackgroundStyle.Color1 == current.ScreenStyle.BackgroundStyle.Color1;
-
-    private static bool CanReuse(BoxBounds previous, BoxBounds current)
-        => PreviewPane.CanReuse(previous.OuterBounds, current.OuterBounds)
-        && PreviewPane.CanReuse(previous.PaddingBounds, current.PaddingBounds)
-        && PreviewPane.CanReuse(previous.ContentBounds, current.ContentBounds);
-
-    private static bool CanReuse(RectangleInfo previous, RectangleInfo current)
-        => previous.X == current.X
-        && previous.Y == current.Y
-        && previous.Width == current.Width
-        && previous.Height == current.Height;
-
-    private static bool CanReuse(BorderStyle previous, BorderStyle current)
-        => previous.Color == current.Color
-        && previous.Left == current.Left
-        && previous.Top == current.Top
-        && previous.Right == current.Right
-        && previous.Bottom == current.Bottom
-        && previous.Depth == current.Depth;
-
-    private void RemoveScreenSlot(ScreenSlot? slot)
-    {
-        if (slot is null)
-        {
-            return;
-        }
-
-        this.ScreensCanvas.Children.Remove(slot.BezelImage);
-        if (slot.PlaceholderRectangle is not null)
-        {
-            this.ScreensCanvas.Children.Remove(slot.PlaceholderRectangle);
-        }
-
-        this.ScreensCanvas.Children.Remove(slot.ContentImage);
-    }
-
-    /// <summary>
-    /// Creates the visual elements for a single screen - a bezel (rendered the same way as the
-    /// host's own outer border, see <see cref="DrawingHelper.RenderBorder"/>), a placeholder
-    /// fill shown until the real screenshot arrives, and a content <see cref="Image"/> that
-    /// starts either blank or showing this screen's last real screenshot, blurred (see
-    /// <see cref="GetHistoricalPlaceholder"/>), until it's populated for real via
-    /// <see cref="SetScreenshot"/>. The bezel and placeholder/content don't overlap - a screen's
-    /// border ring and its content area are disjoint (see <see cref="Models.Drawing.BoxBounds"/>)
-    /// - so paint order between them doesn't matter.
-    /// </summary>
-    private ScreenSlot CreateScreenSlot(ScreenLayout screenLayout, double scale)
-    {
-        var screenBounds = screenLayout.ScreenBounds;
-
-        Image bezelImage;
-        using (var bezelBitmap = DrawingHelper.RenderBorder(
-            screenBounds.MoveTo(new PointInfo(0, 0)), screenLayout.ScreenStyle))
-        {
-            bezelImage = new Image
-            {
-                Source = PreviewPane.ToBitmapImage(bezelBitmap),
-                Stretch = Stretch.Fill,
-            };
-            PreviewPane.PositionElement(bezelImage, screenBounds.OuterBounds, scale);
-            this.ScreensCanvas.Children.Add(bezelImage);
-        }
-
-        Microsoft.UI.Xaml.Shapes.Rectangle? placeholder = null;
-        var placeholderColor = screenLayout.ScreenStyle.BackgroundStyle.Color1;
-        if (placeholderColor is not null)
-        {
-            placeholder = new Microsoft.UI.Xaml.Shapes.Rectangle
-            {
-                Fill = new SolidColorBrush(PreviewPane.ToWindowsColor(placeholderColor.Value)),
-            };
-            PreviewPane.PositionElement(placeholder, screenBounds.PaddingBounds, scale);
-            this.ScreensCanvas.Children.Add(placeholder);
-        }
-
-        var contentImage = new Image
-        {
-            Source = this.GetHistoricalPlaceholder(screenLayout.ScreenInfo),
-            Stretch = Stretch.Fill,
-        };
-        PreviewPane.PositionElement(contentImage, screenBounds.ContentBounds, scale);
-        this.ScreensCanvas.Children.Add(contentImage);
-
-        return new ScreenSlot(screenLayout, bezelImage, placeholder, contentImage);
-    }
-
-    private double GetRasterizationScale()
-        => this.XamlRoot?.RasterizationScale ?? 1.0;
-
-    private static void PositionElement(FrameworkElement element, RectangleInfo bounds, double scale)
-    {
-        element.Width = (double)bounds.Width / scale;
-        element.Height = (double)bounds.Height / scale;
-        Canvas.SetLeft(element, (double)bounds.X / scale);
-        Canvas.SetTop(element, (double)bounds.Y / scale);
-    }
-
-    /// <summary>
-    /// Copies <paramref name="bitmap"/>'s pixels directly into a <see cref="WriteableBitmap"/>,
-    /// rather than round-tripping through a PNG encode (GDI+) + decode (WinUI) - every screen
-    /// slot creates two of these (bezel and content) per activation, so the encode/decode cost
-    /// that was previously paid once or twice for a whole composited canvas was instead being
-    /// paid per screen, which added up. This works as a straight byte copy because every
-    /// bitmap this control receives is already <see cref="PixelFormat.Format32bppPArgb"/>
-    /// (see <see cref="DrawingHelper.RenderBorder"/>/<c>IScreenshotCaptureProvider</c>
-    /// implementations), which is byte-for-byte the same layout as
-    /// <see cref="WriteableBitmap"/>'s own pixel buffer (BGRA8, premultiplied).
-    /// </summary>
-    private static WriteableBitmap ToBitmapImage(Bitmap bitmap)
-    {
-        var writeableBitmap = new WriteableBitmap(bitmap.Width, bitmap.Height);
-
-        var bounds = new Rectangle(0, 0, bitmap.Width, bitmap.Height);
-        var bitmapData = bitmap.LockBits(bounds, ImageLockMode.ReadOnly, PixelFormat.Format32bppPArgb);
-        try
-        {
-            var byteCount = bitmapData.Stride * bitmapData.Height;
-            var buffer = new byte[byteCount];
-            Marshal.Copy(bitmapData.Scan0, buffer, 0, byteCount);
-
-            using var pixelStream = writeableBitmap.PixelBuffer.AsStream();
-            pixelStream.Write(buffer, 0, byteCount);
-        }
-        finally
-        {
-            bitmap.UnlockBits(bitmapData);
-        }
-
-        writeableBitmap.Invalidate();
-        return writeableBitmap;
-    }
-
-    private static Windows.UI.Color ToWindowsColor(Color color)
-        => Windows.UI.Color.FromArgb(color.A, color.R, color.G, color.B);
-
-    private void PreviewPane_PointerPressed(object sender, PointerRoutedEventArgs e)
-    {
-        if (!e.Pointer.PointerDeviceType.Equals(PointerDeviceType.Mouse))
-        {
-            // not a mouse click
-            return;
-        }
-
-        var pointerPoint = e.GetCurrentPoint(this);
-        if (pointerPoint.Properties.IsRightButtonPressed)
-        {
-            this.Cancel?.Invoke(this, EventArgs.Empty);
-            return;
-        }
-
-        if (!pointerPoint.Properties.IsLeftButtonPressed)
-        {
-            return;
-        }
-
-        var layout = this.Layout;
-        if (layout is null)
-        {
-            return;
-        }
-
-        // the pointer position is in DIPs, relative to this control - scale it up to the
-        // same physical-pixel space that Layout's bounds are expressed in
-        var scale = this.GetRasterizationScale();
-        var pointerLocation = new PointInfo((decimal)pointerPoint.Position.X, (decimal)pointerPoint.Position.Y)
-            .Scale((decimal)scale);
-
-        // work out which screenshot was clicked, keeping the owning device too
-        var clickedEntry = layout.CanvasLayout.DeviceLayouts
-            .SelectMany(deviceLayout => deviceLayout.ScreenLayouts.Select(
-                screenLayout => (DeviceLayout: deviceLayout, ScreenLayout: screenLayout)))
-            .SingleOrDefault(
-                entry => entry.ScreenLayout.ScreenBounds.OuterBounds.Contains(pointerLocation));
-        if (clickedEntry.ScreenLayout is null)
-        {
-            return;
-        }
-
-        // scale up the click onto the physical screen - the aspect ratio of the screenshot
-        // might be distorted compared to the physical screen due to the borders around the
-        // screenshot, so we need to work out the target location on the physical screen first
-        var clickedDisplayArea = clickedEntry.ScreenLayout.ScreenInfo.DisplayArea;
-        var clickedLocation = pointerLocation
-            .Stretch(
-                source: clickedEntry.ScreenLayout.ScreenBounds.ContentBounds,
-                target: clickedDisplayArea)
-            .Clamp(
-                new(
-                    x: clickedDisplayArea.X + 1,
-                    y: clickedDisplayArea.Y + 1,
-                    width: clickedDisplayArea.Width - 1,
-                    height: clickedDisplayArea.Height - 1))
-            .Truncate();
-
-        this.NavigateTo?.Invoke(
-            this, new NavigateToEventArgs(clickedEntry.DeviceLayout.DeviceInfo, clickedLocation));
-    }
-
-    /// <summary>
-    /// Maps 1-9/numpad/P/Left/Right/Home/End to <see cref="NavigateTo"/> and Escape to
-    /// <see cref="Cancel"/>, entirely from <see cref="Layout"/> and <see cref="ActiveScreen"/> -
-    /// no live desktop/cursor state needed. Screen numbering and Home/End use the flattened
-    /// device/screen order (<see cref="GetOrderedScreens"/>), which matches
-    /// <c>ScreenHelper.GetAllScreens()</c>'s order since <c>LayoutHelper.GetPreviewLayout</c>
-    /// builds <see cref="Layout"/> from that same list via order-preserving projections.
-    /// </summary>
-    private void PreviewPane_PreviewKeyDown(object sender, KeyRoutedEventArgs e)
-    {
-        if (e.Key == VirtualKey.Escape)
-        {
-            this.Cancel?.Invoke(this, EventArgs.Empty);
-            return;
-        }
-
-        var layout = this.Layout;
-        var activeScreen = this.ActiveScreen;
-        if (layout is null || activeScreen is null)
-        {
-            return;
-        }
-
-        var orderedScreens = this.GetOrderedScreens(layout);
-        if (orderedScreens.Count == 0)
-        {
-            return;
-        }
-
-        var currentIndex = orderedScreens.FindIndex(
-            entry => entry.ScreenLayout.ScreenInfo.Equals(activeScreen));
-
-        var target = e.Key switch
-        {
-            >= VirtualKey.Number1 and <= VirtualKey.Number9 =>
-                PreviewPane.ElementAtOrDefault(orderedScreens, (e.Key - VirtualKey.Number0) - 1),
-            >= VirtualKey.NumberPad1 and <= VirtualKey.NumberPad9 =>
-                PreviewPane.ElementAtOrDefault(orderedScreens, (e.Key - VirtualKey.NumberPad0) - 1),
-            VirtualKey.P =>
-                orderedScreens.SingleOrDefault(entry => entry.ScreenLayout.ScreenInfo.Primary),
-            VirtualKey.Left when currentIndex >= 0 =>
-                orderedScreens[(currentIndex - 1 + orderedScreens.Count) % orderedScreens.Count],
-            VirtualKey.Right when currentIndex >= 0 =>
-                orderedScreens[(currentIndex + 1) % orderedScreens.Count],
-            VirtualKey.Home => orderedScreens[0],
-            VirtualKey.End => orderedScreens[^1],
-            _ => default,
-        };
-
-        if (target.ScreenLayout is null)
-        {
-            return;
-        }
-
-        this.NavigateTo?.Invoke(
-            this,
-            new NavigateToEventArgs(
-                target.DeviceLayout.DeviceInfo, target.ScreenLayout.ScreenInfo.DisplayArea.Midpoint));
-    }
-
-    private List<(DeviceLayout DeviceLayout, ScreenLayout ScreenLayout)> GetOrderedScreens(PreviewLayout layout)
-        => layout.CanvasLayout.DeviceLayouts
-            .SelectMany(deviceLayout => deviceLayout.ScreenLayouts.Select(
-                screenLayout => (DeviceLayout: deviceLayout, ScreenLayout: screenLayout)))
-            .ToList();
-
-    private static (DeviceLayout DeviceLayout, ScreenLayout ScreenLayout) ElementAtOrDefault(
-        List<(DeviceLayout DeviceLayout, ScreenLayout ScreenLayout)> screens, int index)
-        => (index >= 0 && index < screens.Count) ? screens[index] : default;
 }
